// src/app/user-management/model/answer-dto.ts
export interface AnswerDto {
  section?: string;
  situation?: { key: string; value: string };
  doctoratSubject?: { key: string; value: string };
  laboName?: { key: string; value: string };
  comment?: { key: string; value: string };
  employment?: { key: string; value: string };
  contract?: { key: string; value: string };
  function_name?: { key: string; value: string };
  salary?: { key: string; value: string };
  company?: { key: string; value: string };
}